+++ src/rma/sampling/random.cpp
@@ -29,7 +29,7 @@ Random::Random(
     const ssize_t x_dims = x.ndim - 1;
     const ssize_t y_dims = y.ndim - 1;
 
-    if (structure.size() != x_dims + y_dims) throw std::invalid_argument("Structure is not compatible with the input data.");
+    if (static_cast<ssize_t>(structure.size()) != x_dims + y_dims) throw std::invalid_argument("Structure is not compatible with the input data.");
     if (x.shape[0] != y.shape[0]) throw std::invalid_argument("`x` and `y` first dimension must have same size.");
 
     ssize_t total_motifs = 1;
